+++ lib/ecto_shorts/query_builder.ex
@@ -1,20 +1,34 @@
 defmodule EctoShorts.QueryBuilder do
-  @moduledoc "Behaviour for query building from filter tuples"
+  @moduledoc """
+  Specifies the query builder API required from adapters.
+  """
 
-  @type filter_tuple :: {filter_type :: atom, value :: any}
-  @type accumulator_query :: Ecto.Query.t
+  @type adapter :: module()
+  @type filter_key :: atom()
+  @type filter_value :: any()
+  @type query :: Ecto.Query.t()
+  @type queryable :: Ecto.Queryable.t()
 
-  @doc "Adds to accumulator query with filter_type and value"
-  @callback create_schema_filter(filter_tuple, accumulator_query) :: Ecto.Query.t
+  @doc """
+  Adds an expression to the query given a filter key and value.
+  """
+  @callback create_schema_filter(query(), filter_key(), filter_value()) :: query()
 
-  @spec create_schema_filter(module, filter_tuple, accumulator_query) :: Ecto.Query.t
-  def create_schema_filter(builder, filter_tuple, query) do
-    builder.create_schema_filter(filter_tuple, query)
+  @doc """
+  Invokes the callback function `c:EctoShorts.QueryBuilder.create_schema_filter/3`
+  on the given adapter.
+
+  ### Examples
+
+      iex> EctoShorts.QueryBuilder.create_schema_filter(
+      ...>   EctoShorts.QueryBuilder.Common,
+      ...>   EctoShorts.Support.Schemas.Comment,
+      ...>   :first,
+      ...>   1_000
+      ...> )
+  """
+  @spec create_schema_filter(adapter(), query() | queryable(), filter_key(), filter_value()) :: query()
+  def create_schema_filter(adapter, query, filter_key, filter_value) do
+    adapter.create_schema_filter(query, filter_key, filter_value)
   end
-
-  @spec query_schema(Ecto.Queryable.t) :: Ecto.Queryable.t()
-  @doc "Pulls the schema from a query"
-  def query_schema(%{from: %{source: {_, schema}}}), do: query_schema(schema)
-  def query_schema(%{from: %{query: %{from: {_, schema}}}}), do: schema
-  def query_schema(query), do: query
 end